Avoid running rules more than once.
Until now rules had dependencies which were always executed first.

A ...
B ...
C { dependency => [A,B] }
D { dependency => [B] }

We'd run something like:

```
A
B
 A -> dependency of C
 B
C
 B -> dependency of D
D
```

Now we ensure that we run each rule only once; once we've run the
rule we record that.  No need to run the rules again, so with this
change we'd run:

```
A
B
C -> We know we've already run A,B
D -> We know we've already run B
```