Allow modules access to (global) configuration structure
At the moment we only have one command-line option, `-verbose`,
which is used to control how noisy to be when running a rule-file.

However it would be nice if modules could pay attention to that
flag as well as our execution-engine.  To allow that we have
now defined a structure to hold all global options, and made
it available to our plugin/modules.

This closes #22.