WIP: Move towards an AST-walking interpreter
This pull-request, when complete, will rework our internals such
that we have a tree-walking AST interpreter.  This means that
variable processing, string expansion, and file/rule-inclusion
will happen at runtime rather than at parse-time.

Currently this is a work in-progress.

This initial commit lays the foundation, and updates the parser
such that we return a program consisting of various node-types:

* Assignments.
* Inclusions.
* Rules.

Things are not yet executed, but we dump them to the console
for sanity:

        frodo ~/Repos/github.com/skx/marionette $ ./marionette input.txt
        Assignment: foo -> bar
        Assignment: today -> Tue 04 Jan 2022 06:07:48 PM EET
        RULE: directory ..
        RULE: directory ..
        RULE: shell ..
        RULE: file ..
        RULE: file ..
        RULE: file ..
        Assignment: fmt -> +%Y
        RULE: file ..
        RULE: file ..
        RULE: shell ..
        RULE: shell ..
        RULE: shell ..
        RULE: link ..
        RULE: git ..
        RULE: file ..
        RULE: edit ..
        Assignment: user -> server
        RULE: shell ..