Allow assignments to be conditional.
This pull-request overhauls how the conditional-keys were handled
for rule-execution:

* In the past we stored the conditions in the map of parameters
  * Under keys "unless" or "if".
* Now we store them in distinct places within our Rule-node
  * Just like we did for file-inclusion.

This means that we have two objects which support conditional
execution and they are both handled the same way:

* Inclusions
* Rules

Once that was done I added the same struct-members to the
assignment object, and updated the parser to populate them.

That means all three of our nodes support conditional "stuff", and
it is handled in the same way.

This closes #72