Simplify the expansion of variables/commands.
I'm trying to simplify the executor package, and an obvious
thing to remove is the string-expansion and command-execution.

Since expansion largely revolves around the replacement of
variables we can shift that stuff, logically, to the environment
package.

Sure executing commands is a bit weird there, but otherwise it
makes sense and is clean and readable.