Added SQL-module
The SQL module allows you to connect to MySQL, or SQLite, and run
arbitrary SQL statement(s).

You must specify the driver to use, and the DSN, as well as provide
either an SQLite statement or a file containing a list of statements.

Given that we always run the same SQL it should be noted that care
will be needed to avoid making changes.  Much like the shell-module
the SQL module will *always* result in a "change" result.

This closes #95.