Define new AST structure.
This commit defines new AST types which we're going to use in the
future:

* Backtick
* Boolean
* Number
* String

At the moment these are not used, the parser needs to be updated
to process these as values for block-variables - which will necessitate
updating the `Rule` and `Assign` types.

For the moment this is laying the groundwork.