Update our object-types.
This commit removes the last use of the token.Token objects, and
updates our parser to use a consistent set of types for the Objects
we need to use.

Rather than using the general `Node` we're usign the more specific
`Object` type appropriately, and we're using a `Funcall` explicitly
for our conditional execution.