Group packages to install/uninstall in a single command
This change means that the following input:

        package {
          package => [ "foo", "bar" ],
          state => "installed"
        }

Will now run:

        apt-get install foo bar

Rather than result in:

        apt-get install foo
        apt-get install bar

There might be side-effects here, if two incompability packages
are installed at the same time, but that should be "obvious" to the
user.

Also added extra logging to highlight the change, when running
with `--debug`

This closes #110.